Adds CLI Class
This commit adds a bare bones CLI class to get our first end-to-end test case
working. Right now it's doing a little too much nastiness with OptionParser.
Eventually all of that should be abstracted, especially since OptionParser
doesn't give a good way to pass in streams for stdin and stdout.

The tradeoff I made here was to do a little working around OptionParser to save
myself adding a third-party depedency.