Refactor Parsing And Calculation Logic
This was a big change, and lingered longer in red than I had hoped, but the
basic idea was to separate the lexing of input, the calculation, and the
printing logic. Tests are all green now, though there on no tests for the new
logger, lexer, or calculator classes. The plan is to flesh out those tests and
move tests over as I go to add features.

The hope is that now accomidating the rest of the operators will be quick
changes.

As a side note, the calculation logic should work with any proc that takes two
inputs.  I thought about making the number of inputs variable, but I actually
have a bug to two living in that logic that I should fix first.