Implement The Rest Of The Operators and Operand Types
This commit adds the rest of the operators and handles Zero Division. I'm not in
love with the way that works since my calculator was so clean. It will still
work with nearly any function that takes two inputs and returns a value, however
it's leaky and hinting at an abstraction that I missed.

This also lexes and converts all numbers to floats in order to avoid integer
division issues.